# Performance improvements to WKB unmarshalling