as discussed in https://github.com/peterstace/simplefeatures/issues/419 this PR improves the Unmarshalling performance of `WKB` by avoiding the `binary.ByteOrder` interface.

this PR itself is purely cosmetic, so I didn't add/change any testing code. it relies on the existing tests to confirm it's a no-op refactor.

I will follow up with a subsequent PR to commit the benchmarking script, that benchmark code is actually more complex than this 😆 so I decided to decouple it from the 'actual work' here.

Overall it shows a marked improvement with the exception of `201ns ± 1%     214ns ± 6%   +6.49%` which is so fast anyway that it's probably just a benchmark fart.

```bash
benchstat -delta-test ttest master.bench pr.bench
name           old time/op    new time/op    delta
WKBParse/0-8      316ns ± 0%     303ns ± 1%   -4.18%  (p=0.000 n=5+5)
WKBParse/1-8      322ns ± 0%     311ns ± 2%   -3.32%  (p=0.005 n=5+5)
WKBParse/2-8      387ns ± 4%     364ns ± 0%   -5.91%  (p=0.008 n=5+5)
WKBParse/3-8      390ns ± 3%     365ns ± 0%   -6.55%  (p=0.002 n=5+5)
WKBParse/4-8      491ns ± 5%     458ns ± 0%   -6.79%  (p=0.006 n=5+5)
WKBParse/5-8      355ns ± 6%     335ns ± 4%     ~     (p=0.058 n=5+5)
WKBParse/6-8      410ns ± 3%     384ns ± 0%   -6.29%  (p=0.002 n=5+4)
WKBParse/7-8      412ns ± 3%     387ns ± 1%   -6.06%  (p=0.001 n=5+5)
WKBParse/8-8      513ns ± 3%     485ns ± 0%   -5.42%  (p=0.005 n=5+5)
WKBParse/9-8      230ns ± 3%     216ns ± 0%   -5.81%  (p=0.004 n=5+4)
WKBParse/10-8     232ns ± 2%     217ns ± 0%   -6.63%  (p=0.000 n=5+5)
WKBParse/11-8     230ns ± 3%     218ns ± 2%   -4.95%  (p=0.008 n=5+5)
WKBParse/12-8     233ns ± 3%     224ns ± 6%     ~     (p=0.056 n=5+5)
WKBParse/13-8     544ns ± 1%     539ns ± 0%     ~     (p=0.056 n=5+5)
WKBParse/14-8     679ns ± 2%     669ns ± 1%     ~     (p=0.053 n=5+5)
WKBParse/15-8     682ns ± 3%     669ns ± 0%     ~     (p=0.101 n=5+5)
WKBParse/16-8     845ns ± 3%     833ns ± 1%     ~     (p=0.276 n=5+5)
WKBParse/17-8     702ns ± 1%     668ns ± 0%   -4.83%  (p=0.000 n=4+5)
WKBParse/18-8    1.05µs ±13%    0.91µs ± 0%     ~     (p=0.065 n=5+5)
WKBParse/19-8     939ns ± 4%     938ns ± 4%     ~     (p=0.977 n=5+5)
WKBParse/20-8    1.09µs ± 1%    1.14µs ± 8%     ~     (p=0.128 n=5+5)
WKBParse/21-8     201ns ± 1%     214ns ± 6%   +6.49%  (p=0.030 n=5+5)
WKBParse/22-8     205ns ± 2%     224ns ±17%     ~     (p=0.149 n=5+5)
WKBParse/23-8     204ns ± 2%     202ns ± 2%     ~     (p=0.223 n=5+5)
WKBParse/24-8     212ns ± 9%     207ns ± 4%     ~     (p=0.483 n=5+5)
WKBParse/25-8    1.78µs ± 2%    1.76µs ± 3%     ~     (p=0.349 n=5+5)
WKBParse/26-8    2.27µs ± 1%    2.26µs ± 4%     ~     (p=0.660 n=5+5)
WKBParse/27-8    2.32µs ± 1%    2.28µs ± 3%     ~     (p=0.182 n=4+5)
WKBParse/28-8    3.09µs ±12%    2.81µs ± 0%   -9.10%  (p=0.044 n=5+4)
WKBParse/29-8     209ns ± 3%     211ns ± 6%     ~     (p=0.640 n=5+5)
WKBParse/30-8     210ns ± 3%     207ns ± 3%     ~     (p=0.342 n=5+5)
WKBParse/31-8     213ns ± 3%     206ns ± 2%   -3.15%  (p=0.039 n=5+5)
WKBParse/32-8     231ns ±14%     209ns ± 4%     ~     (p=0.125 n=5+5)
WKBParse/33-8     566ns ±14%     528ns ± 3%     ~     (p=0.138 n=5+5)
WKBParse/34-8     651ns ± 5%     612ns ± 1%   -6.00%  (p=0.012 n=5+5)
WKBParse/35-8     649ns ± 2%     611ns ± 1%   -5.79%  (p=0.000 n=5+5)
WKBParse/36-8     768ns ±17%     679ns ± 1%     ~     (p=0.056 n=5+5)
WKBParse/37-8     942ns ±18%     817ns ± 1%  -13.35%  (p=0.047 n=5+5)
WKBParse/38-8    1.04µs ± 5%    0.99µs ± 0%   -5.28%  (p=0.020 n=5+4)
WKBParse/39-8    1.01µs ± 0%    0.99µs ± 0%   -1.74%  (p=0.000 n=5+5)
WKBParse/40-8    1.17µs ± 1%    1.14µs ± 0%   -2.49%  (p=0.000 n=5+5)
WKBParse/41-8     204ns ± 1%     200ns ± 0%   -1.66%  (p=0.002 n=5+5)
WKBParse/42-8     205ns ± 0%     202ns ± 1%   -1.60%  (p=0.000 n=5+5)
WKBParse/43-8     225ns ±17%     200ns ± 0%     ~     (p=0.061 n=5+5)
WKBParse/44-8     213ns ± 2%     201ns ± 0%   -5.48%  (p=0.001 n=5+4)
WKBParse/45-8    1.00µs ± 3%    0.92µs ± 0%   -8.68%  (p=0.001 n=5+4)
WKBParse/46-8    1.24µs ±10%    1.12µs ± 2%   -9.69%  (p=0.021 n=5+5)
WKBParse/47-8    1.36µs ±16%    1.11µs ± 0%  -18.29%  (p=0.042 n=5+5)
WKBParse/48-8    1.39µs ±14%    1.29µs ± 1%     ~     (p=0.127 n=5+5)
WKBParse/49-8    1.49µs ± 4%    1.39µs ± 1%   -6.47%  (p=0.005 n=5+5)
WKBParse/50-8    2.08µs ± 9%    1.88µs ±11%     ~     (p=0.057 n=5+5)
WKBParse/51-8    2.12µs ±14%    1.81µs ± 5%  -14.51%  (p=0.022 n=5+5)
WKBParse/52-8    2.32µs ±12%    2.11µs ± 4%   -9.02%  (p=0.039 n=5+5)
WKBParse/53-8     230ns ±18%     198ns ± 2%     ~     (p=0.056 n=5+5)
WKBParse/54-8     219ns ± 7%     199ns ± 1%   -8.91%  (p=0.015 n=5+5)
WKBParse/55-8     213ns ± 9%     199ns ± 2%   -6.61%  (p=0.048 n=5+5)
WKBParse/56-8     207ns ± 1%     199ns ± 0%   -4.07%  (p=0.000 n=5+4)
WKBParse/57-8    2.34µs ± 5%    2.26µs ± 0%   -3.40%  (p=0.048 n=5+5)
WKBParse/58-8    3.02µs ± 3%    2.78µs ± 0%   -8.23%  (p=0.001 n=5+5)
WKBParse/59-8    3.02µs ± 3%    2.78µs ± 0%   -7.77%  (p=0.002 n=5+5)
WKBParse/60-8    3.63µs ± 3%    3.38µs ± 0%   -6.79%  (p=0.004 n=5+5)
WKBParse/61-8     207ns ± 5%     203ns ± 3%     ~     (p=0.334 n=5+5)
WKBParse/62-8     203ns ± 2%     202ns ± 1%     ~     (p=0.296 n=5+5)
WKBParse/63-8     202ns ± 0%     204ns ± 2%     ~     (p=0.242 n=4+5)
WKBParse/64-8     207ns ± 4%     204ns ± 2%     ~     (p=0.275 n=5+5)
WKBParse/65-8     971ns ± 4%     947ns ± 6%     ~     (p=0.329 n=5+5)
WKBParse/66-8    1.15µs ± 2%    1.10µs ± 0%   -4.34%  (p=0.002 n=5+5)
WKBParse/67-8    1.13µs ± 1%    1.10µs ± 0%   -2.57%  (p=0.000 n=5+5)
WKBParse/68-8    1.30µs ± 2%    1.26µs ± 0%   -3.19%  (p=0.005 n=5+5)
```

closes https://github.com/peterstace/simplefeatures/issues/419