@peterstace 2022-07-08
Even without the benchmarks from https://github.com/peterstace/simplefeatures/pull/422 not in a state to merge as it is, I'm convinced enough to take this change on its own. Thanks for the contribution!